Task_1_rag/requirements.txt
streamlit==1.35.0
supabase==2.28.0
PyMuPDF==1.27.2.2
requests>=2.31.0
python-dotenv>=1.0.1
